cleaned it up a bit
